.github/workflows/test.yml: Require that index.js is up to date
Since this action action uses `dist/index.js` as the backing code
instead of the files under `src/`, perform the check that
`dist/index.js` is valid before running any of the other tests.

This is meant to catch situations in which changes are committed
without the `index.js` file being updated. In that case the results of
the other tests are useless as they would have used the old code
instead of the newly committed one.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>