action.yml: Stop defaulting to dmd-latest as the D compiler
There is already code in the action that picks between dmd-latest and
ldc-latest based on the CPU architecture when a D compiler is not
specified. Having the default value in action.yml prevented this code
from ever running. This leads to users receiving an error message and
having to specify `dc: ldc` manually.

Bug: https://github.com/dlang-community/setup-dlang/issues/78

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>