Add unittests written in typescript
Currently all of the tests are written as a github workflow. This
makes them very inconvenient to run as one would have to be triggering
the online CI every times they make a change to the code. On top of
this the tests can only verify the final product of the action, if
changes are made somewhere deep in the source code and they break
something it will not be immediately clear what is broken.

For this reason I've written tests in plain typescript that are much
easier to be run but still check all the platforms (linux, windows,
and macos) on x86_64 and arm64. These tests are covering a big part of
the source code, mostly the logic bits in addition to mocking network
responses for the code that contacts github.com or dlang.org to
determine the versions that needs to be installed. This should make
future work on the action way easier.

This doesn't make the old tests obsolete as some parts of the code are
not easily tested, like fetching release archives and properly
extracting them.

Add an action that will run the unittests on each system (linux,
macos, windows) to assure that any person can run the tests on their
system without failures.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>