Set the DMD environment variable for all compilers
This action sets up DC to point to the compiler. It is not much more
work to also set DMD to point to the dmd wrapper of the compiler. dmd
is its own wrapper, ldmd2 comes by default with ldc2. The only one
left out is gdc.

The action allowed installing gdc, either alone or with gdmd and
setting DC to point to one of them. This required the user to specify
which what they wanted to set DC to. This is pointlessly complex as
gdmd is just a script file that can be easily downloaded. The previous
decision to do this separately was because of the installation method
of the gdmd script, it defaulting to installing it through `apt`,
which would likely lead to many more useless packages being
installed. This is too much work for just a script file so it will now
always be installed by fetching it from the upstream repo.  See
https://github.com/dlang-community/setup-dlang/pull/77#discussion_r1622610721
for a discussion around this topic.

This has as consequences that the `gdmd` and `gdmd-<version>` inputs
to `compiler` become invalid and that the `gdmd_sha` inputs becomes
mandatory. To accommodate for the second issue the action now picks a
default value for `gdmd_sha`.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>

Add CI for gdc and gdmd

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>