.github/actions/verify-d-compiler: Perform the tests in separate dirs
Avoid the possibility of mistakenly running a previously generated
hello executable by performing all the steps in separate directories.

This has shown an issue with the `-shared` test, it never produced an
executable so invoking the program was useless as it run the one that
was generated in the previous step. For this reason in that test
perform only compilation and no longer run any binary.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>